requirements.txt
Flask
pymongo
python-dotenv
langchain
langchain-openai
langchain-mongodb
PyPDF2
tiktoken

